MB-11405 Reduce CPU use in active system
Instead of having all the threads snooze for the minimum
snooze time across all task sets, have only 1 thread per
task set snooze while others sleep for longer duration.
This not only uses lower cpu in an active system with 5+ buckets
but also reduces lock contention in smaller configurations.

Reviewed-on: http://review.couchbase.org/38620
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>

+++ src/executorthread.h
@@ -56,8 +56,9 @@ class ExecutorThread {
     friend class ExecutorPool;
 public:
 
-    ExecutorThread(ExecutorPool *m, size_t startingQueue, const std::string nm)
-        : manager(m), startIndex(startingQueue), name(nm),
+    ExecutorThread(ExecutorPool *m, int startingQueue,
+                   const std::string nm) : manager(m),
+          startIndex(startingQueue), name(nm),
           state(EXECUTOR_CREATING), taskStart(0),
           currentTask(NULL), curTaskType(NO_TASK_TYPE),
           tasklog(TASK_LOG_SIZE), slowjobs(TASK_LOG_SIZE) {
@@ -107,7 +108,7 @@ private:
 
     cb_thread_t thread;
     ExecutorPool *manager;
-    size_t startIndex;
+    int startIndex;
     const std::string name;
     executor_state_t state;
